+++ src/prommis/nanofiltration/tests/test_diafiltration_solute_properties.py
@@ -14,7 +14,7 @@ from idaes.core import FlowsheetBlock
 
 import pytest
 
-from prommis.nanofiltration.diafiltration_solute_properties import (
+from prommis.nanofiltration.multi_component_diafiltration_solute_properties import (
     SoluteParameter,
 )
